@@ -6,7 +6,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"io/ioutil"
-	"log"
+	"log/slog"
 	"net/http"
 	"strings"
 
@@ -15,6 +15,10 @@ import (
 )
 
 func processSbomInsightsData(h *Messaging, insights InsightsData, v string, apiClient graphql.Client, resource ResourceDestination) ([]LagoonFact, string, error) {
+
+	source := fmt.Sprintf("insights:sbom:%s", resource.Service)
+	logger := slog.With("ProjectName", resource.Project, "EnvironmentName", resource.Environment, "Source", source)
+
 	if insights.InsightsType != Sbom {
 		return []LagoonFact{}, "", nil
 	}
@@ -59,10 +63,26 @@ func processSbomInsightsData(h *Messaging, insights InsightsData, v string, apiC
 	if apiErr != nil {
 		return nil, "", apiErr
 	}
-	source := fmt.Sprintf("insights:sbom:%s", resource.Service)
+
+	// we process the SBOM here
+
+	if h.ProblemsFromSBOM == true {
+		isAlive, err := IsTrivyServerIsAlive(h.TrivyServerEndpoint)
+		if err != nil {
+			return nil, "", fmt.Errorf("trivy server not alive: %v", err.Error())
+		} else {
+			logger.Debug("Trivy is reachable")
+		}
+		if isAlive {
+			err = SbomToProblems(apiClient, h.TrivyServerEndpoint, "/tmp/", environment.Id, resource.Service, *bom)
+		}
+		if err != nil {
+			return nil, "", err
+		}
+	}
 
 	// Process SBOM into facts
-	facts := processFactsFromSBOM(bom.Components, environment.Id, source)
+	facts := processFactsFromSBOM(logger, bom.Components, environment.Id, source)
 
 	facts, err = KeyFactsFilter(facts)
 	if err != nil {
@@ -73,11 +93,17 @@ func processSbomInsightsData(h *Messaging, insights InsightsData, v string, apiC
 		return nil, "", fmt.Errorf("no facts to process")
 	}
 
-	log.Printf("Successfully decoded SBOM of image %s with %s, found %d for '%s:%s'", bom.Metadata.Component.Name, (*bom.Metadata.Tools)[0].Name, len(*bom.Components), resource.Project, resource.Environment)
+	//log.Printf("Successfully decoded SBOM of image %s with %s, found %d for '%s:%s'", bom.Metadata.Component.Name, (*bom.Metadata.Tools)[0].Name, len(*bom.Components), resource.Project, resource.Environment)
+	logger.Info("Successfully decoded SBOM",
+		"image", bom.Metadata.Component.Name,
+		"fieldName", (*bom.Metadata.Tools)[0].Name,
+		"Length", len(*bom.Components),
+	)
+
 	return facts, source, nil
 }
 
-func processFactsFromSBOM(facts *[]cdx.Component, environmentId int, source string) []LagoonFact {
+func processFactsFromSBOM(logger *slog.Logger, facts *[]cdx.Component, environmentId int, source string) []LagoonFact {
 	var factsInput []LagoonFact
 	if len(*facts) == 0 {
 		return factsInput
@@ -104,9 +130,13 @@ func processFactsFromSBOM(facts *[]cdx.Component, environmentId int, source stri
 			KeyFact:     false,
 			Type:        FactTypeText,
 		}
-		if EnableDebug {
-			log.Println("[DEBUG] processing fact name " + f.Name)
-		}
+		//if EnableDebug {
+		//	log.Println("[DEBUG] processing fact name " + f.Name)
+		//}
+		logger.Debug("Processing fact",
+			"Name", f.Name,
+			"Value", f.Version,
+		)
 		fact, _ = ProcessLagoonFactAgainstRegisteredFilters(fact, f)
 		factsInput = append(factsInput, fact)
 	}
