@@ -3,7 +3,7 @@ package handler
 import (
 	"encoding/json"
 	"fmt"
-	"gopkg.in/yaml.v1"
+	"gopkg.in/yaml.v3"
 	"io/ioutil"
 	"path/filepath"
 	"strings"
@@ -19,7 +19,7 @@ type FactTransformNameValue struct {
 type FactLookupNameValue struct {
 	Name       string `json:"name"`
 	Value      string `json:"value"`
-	ExactMatch bool   `json:"exactMatch"`
+	ExactMatch bool   `json:"exactMatch" yaml:"exactMatch"`
 }
 
 type FactTransform struct {
